auth_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
python-dotenv==1.0.1
PyJWT==2.10.0
passlib==1.7.4
pydantic==2.9.2
email-validator==2.2.0
